refactor: remove pkg from the Go package paths
